Valheim Plus config file updates
Created plaintext Valheim Plus config file for anyone looking for ideas, and updated zipped version to match it (though it's not required, as the server will push updates on its own).

I picked these changes to not be *that* game-breaking for how we play, and primarily focused on quality-of-life improvements like keeping torches lit, increasing build distance slightly, expanding portal usefulness by allowing the passthrough of metals (though I doubt we'll use it), and increasing carry capacity.